cloud_functions/image_handler/requirements.txt
google-cloud-storage
google-cloud-pubsub
google-cloud-bigquery
torch
pillow
transformers
einops
torchvision
gunicorn==22.0.0
Werkzeug==3.0.1
Flask==3.0.0